refactor: keep one version of validator, for skills
