+++ server/lib/schema_web/router.ex
@@ -140,9 +140,8 @@ defmodule SchemaWeb.Router do
 
     post "/enrich", SchemaController, :enrich
     post "/translate", SchemaController, :translate
-    post "/validate", SchemaController, :validate
-    post "/v2/validate", SchemaController, :validate2
-    post "/v2/validate_bundle", SchemaController, :validate2_bundle
+    post "/validate/skill", SchemaController, :validate
+    post "/validate_bundle/skill", SchemaController, :validate_bundle
   end
 
   scope "/schema", SchemaWeb do
